+++ client/Packages/com.beamable/Editor/Modules/Content/ContentIO.cs
@@ -7,6 +7,7 @@ using Beamable.Common.Content.Validation;
 using Beamable.Common.Dependencies;
 using Beamable.Common.Runtime;
 using Beamable.Content;
+using Beamable.Editor.Content.Models;
 using Beamable.Editor.Content.UI;
 using Beamable.Serialization;
 using Core.Platform.SDK;
@@ -32,53 +33,47 @@ namespace Beamable.Editor.Content
 {
 	public class ContentIOAssetProcessor : UnityEditor.AssetModificationProcessor
 	{
+		private static List<ContentDatabaseEntry> _deleteEntries = new List<ContentDatabaseEntry>();
 		private static AssetDeleteResult OnWillDeleteAsset(string assetPath, RemoveAssetOptions options)
 		{
-			if (AssetDatabase.IsValidFolder(assetPath))
+			var db = BeamEditorContext.Default.ServiceScope.GetService<ContentDatabase>();
+			if (db.TryGetContentByPath(assetPath, out var entry))
 			{
-				var subAssets = AssetDatabase.FindAssets("", new string[] { assetPath });
-				foreach (var sub in subAssets)
-				{
-					var subPath = AssetDatabase.GUIDToAssetPath(sub);
-					OnWillDeleteAsset(subPath, options);
-				}
-
-				return AssetDeleteResult.DidNotDelete;
+				_deleteEntries.Add(entry);
 			}
-
-			var asset = AssetDatabase.LoadAssetAtPath<ContentObject>(assetPath);
-			if (asset == null) return AssetDeleteResult.DidNotDelete;
-
-			var method = typeof(ContentIO).GetMethod(nameof(ContentIO.NotifyDeleted),
-				BindingFlags.NonPublic | BindingFlags.Static);
-			var genMethod = method.MakeGenericMethod(asset.GetType());
-			genMethod.Invoke(null, new object[] { asset });
-
+			EditorDebouncer.Debounce("content-delete", () =>
+			{
+				ContentIO.NotifyDeleted(_deleteEntries);
+				_deleteEntries.Clear();
+			});
+			
 			return AssetDeleteResult.DidNotDelete;
+	
 		}
-
-
+	
+	
 		private static string[] OnWillSaveAssets(string[] paths)
 		{
-			foreach (var path in paths)
+			var db = BeamEditorContext.Default.ServiceScope.GetService<ContentDatabase>();
+			if (!db.ContainsAnyContentPaths(paths)) return paths;
+
+			db.Index(); // update, because assets are new!
+			var listOfContent = new List<IContentObject>();
+			for (var i = 0; i < paths.Length; i++)
 			{
-				var asset = AssetDatabase.LoadAssetAtPath<ContentObject>(path);
-				if (asset == null) continue;
-
-				var method = typeof(ContentIO).GetMethod(nameof(ContentIO.NotifyCreated),
-					BindingFlags.NonPublic | BindingFlags.Static);
-				var genMethod = method.MakeGenericMethod(asset.GetType());
-
-				var rootPath = $"{Directories.DATA_DIR}/{asset.ContentType}/";
-				var subPath = path.Substring(rootPath.Length);
-				var qualifiedName = Path.GetFileNameWithoutExtension(subPath.Replace(Path.DirectorySeparatorChar, '.'));
-
-				asset.SetContentName(qualifiedName);
-
-				genMethod.Invoke(null, new object[] { asset });
+				if (!db.TryGetContentByPath(paths[i], out var entry))
+				{
+					continue;
+				}
+	
+				var asset = AssetDatabase.LoadAssetAtPath<ContentObject>(entry.assetPath);
+				asset.SetContentName(entry.contentName);
+				listOfContent.Add(asset);
 			}
-
+			ContentIO.NotifyCreated(listOfContent);
+	
 			return paths;
+			
 		}
 	}
 
@@ -169,6 +164,8 @@ namespace Beamable.Editor.Content
 			return selfIds.SetEquals(otherIds);
 		}
 	}
+	
+	public delegate void IContentEntryDelegate(List<ContentDatabaseEntry> entries);
 
 	/// <summary>
 	/// The purpose of this class is to
@@ -190,13 +187,24 @@ namespace Beamable.Editor.Content
 
 		public event ContentDelegate OnSelectionChanged;
 
-		public static event IContentDelegate OnContentCreated, OnContentDeleted;
+#pragma warning disable CS0067
+		[Obsolete("Do not use. Use " + nameof(OnContentsCreated) + " instead.")]
+		public static event IContentDelegate OnContentCreated;
+		
+		[Obsolete("Do not use. Use " + nameof(OnContentEntryDeleted) + " instead.")]
+		public static event IContentDelegate OnContentDeleted;
+#pragma warning restore CS0067
+
+		public static event IContentBatchDelegate OnContentsCreated;
+		public static event IContentEntryDelegate OnContentEntryDeleted;
+		
 		public static event IContentRenamedDelegate OnContentRenamed;
 		public static Action<string> OnManifestChanged;
 		public static Action<AvailableManifests> OnManifestsListFetched;
 		public static Action<IEnumerable<AvailableManifestModel>> OnArchivedManifestsFetched;
 
 		private ValidationContext ValidationContext => _provider.GetService<ValidationContext>();
+		private ContentDatabase ContentDatabase => _provider.GetService<ContentDatabase>();
 
 		public ContentIO(IDependencyProvider provider, IBeamableRequester requester)
 		{
@@ -204,8 +212,8 @@ namespace Beamable.Editor.Content
 			_requester = requester;
 			Selection.selectionChanged += SelectionChanged;
 
-			OnContentCreated += Internal_HandleContentCreated;
-			OnContentDeleted += Internal_HandleContentDeleted;
+			OnContentsCreated += Internal_HandleContentCreated;
+			OnContentEntryDeleted += Internal_HandleContentDeleted;
 			_contentTypeReflectionCache = BeamEditor.GetReflectionSystem<ContentTypeReflectionCache>();
 
 		}
@@ -217,19 +225,25 @@ namespace Beamable.Editor.Content
 		[RuntimeInitializeOnLoadMethod]
 		private static void Init()
 		{
-			OnContentCreated = null;
-			OnContentDeleted = null;
+			OnContentsCreated = null;
+			OnContentEntryDeleted = null;
 			OnContentRenamed = null;
 		}
 
-		private void Internal_HandleContentCreated(IContentObject content)
+		private void Internal_HandleContentCreated(List<IContentObject> contents)
 		{
-			ValidationContext.AllContent[content.Id] = content;
+			foreach (var content in contents)
+			{
+				ValidationContext.AllContent[content.Id] = content;
+			}
 		}
 
-		private void Internal_HandleContentDeleted(IContentObject content)
+		private void Internal_HandleContentDeleted(List<ContentDatabaseEntry> contents)
 		{
-			ValidationContext.AllContent.Remove(content.Id);
+			foreach (var content in contents)
+			{
+				ValidationContext.AllContent.Remove(content.contentId);
+			}
 		}
 
 
@@ -442,40 +456,23 @@ namespace Beamable.Editor.Content
 
 		private ContentObject Find(ContentObject content)
 		{
-			var assetGuids = BeamableAssetDatabase.FindAssets(
-				content.GetType(),
-				new[] { Directories.DATA_DIR }
-			);
-			foreach (var guid in assetGuids)
+			if (ContentDatabase.TryGetContentById(content.Id, out var entry))
 			{
-				var assetPath = AssetDatabase.GUIDToAssetPath(guid);
-				ContentObject asset = (ContentObject)AssetDatabase.LoadAssetAtPath(assetPath, content.GetType());
-				if (asset.Id == content.Id)
-					return asset;
+				return AssetDatabase.LoadAssetAtPath<ContentObject>(entry.assetPath);
 			}
 
-			throw new Exception("No matching content.");
+			return null;
 		}
 
 		public IEnumerable<TContent> FindAllContent<TContent>(ContentQuery query = null, bool inherit = true)
 			where TContent : ContentObject, new()
 		{
 			if (query == null) query = ContentQuery.Unit;
-			if (!AssetDatabase.IsValidFolder(Directories.DATA_DIR))
-			{
-				Directory.CreateDirectory(Directories.DATA_DIR);
-				yield break; // there is no folder, therefore, no content. Nothing to search for.
-			}
-
-			var assetGuids = BeamableAssetDatabase.FindAssets<TContent>(
-				new[] { Directories.DATA_DIR }
-			);
 			var contentType = _contentTypeReflectionCache.TypeToName(typeof(TContent));
-
-			foreach (var guid in assetGuids)
+			var entries = ContentDatabase.GetContent<TContent>();
+			foreach (var entry in entries)
 			{
-				var path = AssetDatabase.GUIDToAssetPath(guid);
-				var asset = AssetDatabase.LoadAssetAtPath<TContent>(path);
+				var asset = AssetDatabase.LoadAssetAtPath<TContent>(entry.assetPath);
 
 				if (asset == null || !query.Accept(asset))
 					continue;
@@ -483,7 +480,7 @@ namespace Beamable.Editor.Content
 					continue;
 
 
-				var name = Path.GetFileNameWithoutExtension(path);
+				var name = Path.GetFileNameWithoutExtension(entry.assetPath);
 				asset.SetContentName(name);
 				yield return asset;
 			}
@@ -507,27 +504,20 @@ namespace Beamable.Editor.Content
 
 		public IEnumerable<ContentObject> FindAll(ContentQuery query = null)
 		{
-			var contentArray = new List<ContentObject>();
-			var contentIdHashSet = new HashSet<string>();
-			foreach (var contentType in GetContentTypes())
+			foreach (var content in ContentDatabase.LoadAllContent())
 			{
-				foreach (var content in FindAllContentByType(contentType, query))
+				if (query == null || query.Accept(content))
 				{
-					if (!contentIdHashSet.Contains(content.Id))
-					{
-						contentIdHashSet.Add(content.Id);
-						contentArray.Add(content);
-					}
+					yield return content;
 				}
 			}
-
-			return contentArray;
 		}
 
 		public LocalContentManifest BuildLocalManifest()
 		{
 			var localManifest = new LocalContentManifest();
 			ValidationContext.AllContent.Clear();
+			ContentDatabase.Index();
 
 			if (!Directory.Exists(Directories.DATA_DIR))
 			{
@@ -535,40 +525,24 @@ namespace Beamable.Editor.Content
 				return localManifest;
 			}
 
-			foreach (var contentType in _contentTypeReflectionCache.GetContentTypes()) // TODO check hierarchy types.
+			foreach (var entry in ContentDatabase.GetAllContent())
 			{
-				var assetGuids = BeamableAssetDatabase.FindAssets(
-					contentType,
-					new[] { Directories.DATA_DIR }
-				);
-				var typeName = _contentTypeReflectionCache.TypeToName(contentType);
-
-				for (int i = 0; i < assetGuids.Length; i++)
+				var content = AssetDatabase.LoadAssetAtPath<ContentObject>(entry.assetPath);
+				content.SetIdAndVersion(entry.contentId, "");
+		
+				var manifestEntry = new LocalContentManifestEntry
 				{
-					var assetPath = AssetDatabase.GUIDToAssetPath(assetGuids[i]);
-					var rawAsset = AssetDatabase.LoadAssetAtPath(assetPath, typeof(IContentObject));
-					var content = rawAsset as IContentObject;
-
-					var fileName = Path.GetFileNameWithoutExtension(assetPath);
-
-					if (content == null || rawAsset.GetType() != contentType) continue;
-					var contentId = $"{typeName}.{fileName}";
-					content.SetIdAndVersion(contentId, "");
-
-					var entry = new LocalContentManifestEntry
-					{
-						ContentType = contentType,
-						Content = content,
-						AssetPath = assetPath
-					};
-					ValidationContext.AllContent[content.Id] = content;
-					if (!localManifest.Content.ContainsKey(content.Id))
-					{
-						localManifest.Content.Add(content.Id, entry);
-					}
+					ContentType = entry.runtimeType,
+					Content = content,
+					AssetPath = entry.assetPath
+				};
+				ValidationContext.AllContent[content.Id] = content;
+				if (!localManifest.Content.ContainsKey(content.Id))
+				{
+					localManifest.Content.Add(content.Id, manifestEntry);
 				}
 			}
-
+			
 			ContentObject.ValidationContext = ValidationContext;
 			ValidationContext.Initialized = true;
 			return localManifest;
@@ -624,6 +598,7 @@ namespace Beamable.Editor.Content
 			string typeName = ContentObject.GetContentType<TContent>();
 			var dir = $"{Directories.DATA_DIR}/{typeName}";
 			EnsureDefaultAssets<TContent>();
+			var copiedAnydata = false;
 			if (!Directory.Exists(dir))
 			{
 				Directory.CreateDirectory(dir);
@@ -639,10 +614,16 @@ namespace Beamable.Editor.Content
 						var filename = Path.GetFileName(src);
 						var dest = Path.Combine(dir, filename);
 						File.Copy(src, dest, true);
+						copiedAnydata = true;
 					}
 
 					AssetDatabase.ImportAsset(dir, ImportAssetOptions.ImportRecursive);
 				}
+			}
+
+			if (copiedAnydata)
+			{
+				ContentDatabase.Index();
 			}
 		}
 
@@ -650,15 +631,12 @@ namespace Beamable.Editor.Content
 		{
 			foreach (var nextContentType in _contentTypeReflectionCache.GetContentTypes()) // TODO check heirarchy types.
 			{
-				var assetGuids = BeamableAssetDatabase.FindAssets(
-					nextContentType,
-					new[] { Directories.DATA_DIR }
-				);
+				var entries = ContentDatabase.GetContent(contentType);
 
-				foreach (var assetGuid in assetGuids)
+				foreach (var entry in entries)
 				{
-					var assetPath = AssetDatabase.GUIDToAssetPath(assetGuid);
-					var rawAsset = AssetDatabase.LoadAssetAtPath(assetPath, typeof(IContentObject));
+					
+					var rawAsset = AssetDatabase.LoadAssetAtPath(entry.assetPath, typeof(IContentObject));
 					var nextContent = rawAsset as IContentObject;
 
 					if (nextContent == null || rawAsset.GetType() != nextContentType) continue;
@@ -666,7 +644,7 @@ namespace Beamable.Editor.Content
 					if (nextContentType == contentType &&
 						nextContent == content)
 					{
-						return assetPath;
+						return entry.assetPath;
 					}
 				}
 			}
@@ -701,8 +679,6 @@ namespace Beamable.Editor.Content
 				throw new Exception("Addressables was not configured");
 			}
 
-			AssetDatabase.Refresh();
-
 			var filesToMarkAddressable = new List<string>();
 
 			foreach (var src in files)
@@ -716,9 +692,6 @@ namespace Beamable.Editor.Content
 					continue;
 				filesToMarkAddressable.Add(dest);
 			}
-
-			// mark all files as addressable after copy completes...
-			CommitAssetDatabase();
 
 			var addressablesGroup = addressableAssetSettings.FindGroup(BEAMABLE_ASSET_GROUP);
 			if (addressablesGroup == null)
@@ -742,7 +715,6 @@ namespace Beamable.Editor.Content
 
 			addressableAssetSettings.SetDirty(AddressableAssetSettings.ModificationEvent.EntryMoved, addedEntries,
 				true);
-			CommitAssetDatabase();
 		}
 
 		public string GetAvailableFileName(string assetPath, string id, LocalContentManifest localContentManifest)
@@ -776,6 +748,41 @@ namespace Beamable.Editor.Content
 			}
 		}
 
+		public void CreateBatch(IList<Tuple<ContentObject, string>> pathToContent)
+		{
+			var contentList = new List<IContentObject>();
+			try
+			{
+				AssetDatabase.StartAssetEditing();
+				foreach (var (content, assetPath) in pathToContent)
+				{
+					var modifiedAssetPath = assetPath;
+					if (string.IsNullOrEmpty(assetPath))
+					{
+						var newNameAsPath = content.Id.Replace('.', Path.DirectorySeparatorChar);
+						modifiedAssetPath = $"{Directories.DATA_DIR}/{newNameAsPath}.asset";
+					}
+
+					content.name = ""; // force the SO name to be empty. Maintaining two names is too hard.
+					var directory = Path.GetDirectoryName(modifiedAssetPath);
+					Directory.CreateDirectory(directory);
+					AssetDatabase.CreateAsset(content, modifiedAssetPath);
+
+					var modifiedName = Path.GetFileNameWithoutExtension(modifiedAssetPath);
+					content.SetContentName(modifiedName);
+					contentList.Add(content);
+				}
+			}
+			finally
+			{
+				AssetDatabase.StopAssetEditing();
+			}
+
+			ContentDatabase.Index(); // update assets!
+			NotifyCreated(contentList);
+
+		}
+
 		public void Create<TContent>(TContent content, string assetPath = null, bool replace = true, LocalContentManifest localContentManifest = null)
 			where TContent : ContentObject, new()
 		{
@@ -794,6 +801,7 @@ namespace Beamable.Editor.Content
 
 			var modifiedName = Path.GetFileNameWithoutExtension(modifiedAssetPath);
 			content.SetContentName(modifiedName);
+			ContentDatabase.Index(); // update assets!
 			NotifyCreated(content);
 		}
 
@@ -817,20 +825,43 @@ namespace Beamable.Editor.Content
 		public void Delete<TContent>(TContent content)
 			where TContent : ContentObject, new()
 		{
-			var contentName = content.ContentName;
-			var contentSystemType = content.GetType();
-
-			var path = AssetDatabase.GetAssetPath(content);
-			if (string.IsNullOrEmpty(path))
+			if (!ContentDatabase.TryGetContentById(content.Id, out var entry))
 			{
-				var newNameAsPath = contentName.Replace('.', Path.DirectorySeparatorChar);
-				path = $"{Directories.DATA_DIR}/{content.ContentType}/{newNameAsPath}.asset";
+				return;
 			}
+			NotifyDeleted(entry);
 
-			NotifyDeleted(content);
+			AssetDatabase.DeleteAsset(entry.assetPath);
+			File.Delete(entry.assetPath);
+		}
+		
+		public void DeleteBatch(List<IContentObject> contents)
+		{
+			var entries = new List<ContentDatabaseEntry>();
+			foreach (var content in contents)
+			{
+				if (!ContentDatabase.TryGetContentById(content.Id, out var entry))
+				{
+					continue;
+				}
+				entries.Add(entry);
+			}
+			
+			NotifyDeleted(entries);
 
-			AssetDatabase.DeleteAsset(path);
-			File.Delete(path);
+			try
+			{
+				AssetDatabase.StartAssetEditing();
+				foreach (var entry in entries)
+				{
+					AssetDatabase.DeleteAsset(entry.assetPath);
+					File.Delete(entry.assetPath);
+				}
+			}
+			finally
+			{
+				AssetDatabase.StopAssetEditing();
+			}
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -948,13 +979,6 @@ namespace Beamable.Editor.Content
 				ForceReserializeAssetsOptions.ReserializeAssetsAndMetadata);
 		}
 
-		private void CommitAssetDatabase()
-		{
-			// TODO: Revisit this code. It was written this way, because that was the way I could make sure the assets always received the changes
-			AssetDatabase.Refresh();
-			AssetDatabase.SaveAssets();
-			AssetDatabase.Refresh();
-		}
 
 		internal static void NotifyRenamed<TContent>(string oldId, TContent content, string nextAssetPath)
 			where TContent : ContentObject, new()
@@ -965,13 +989,21 @@ namespace Beamable.Editor.Content
 		internal static void NotifyCreated<TContent>(TContent content)
 			where TContent : ContentObject, new()
 		{
-			OnContentCreated?.Invoke(content);
+			OnContentsCreated?.Invoke(new List<IContentObject>{content});
+		}
+		internal static void NotifyCreated(List<IContentObject> content)
+		{
+			OnContentsCreated?.Invoke(content);
 		}
 
-		internal static void NotifyDeleted<TContent>(TContent content)
-			where TContent : ContentObject, new()
+
+		internal static void NotifyDeleted(ContentDatabaseEntry content)
 		{
-			OnContentDeleted?.Invoke(content);
+			OnContentEntryDeleted?.Invoke(new List<ContentDatabaseEntry> {content });
+		}
+		internal static void NotifyDeleted(List<ContentDatabaseEntry> content)
+		{
+			OnContentEntryDeleted?.Invoke(content);
 		}
 
 		private static Promise<ClientManifest> RequestClientManifest(IBeamableRequester requester)